+++ docs/coreComponents/SchemaForm/groupGridHorizontal/index.tsx
@@ -15,9 +15,11 @@ const GroupGridHorizontal = () => {
   return (
     <div>
       <SchemaForm
+        // 👇这是设置给外层的👇
         grid
         rowProps={{ align: 'top' }}
         colProps={{ span: 12 }}
+        // 👆这是设置给外层的👆
         onFinish={onFinish}
         formRef={formRef}
         columns={columns}