+++ CHANGELOG.md
@@ -102,3 +102,8 @@
 ## 0.1.20
 
 - 修改 LinkButton type
+
+## 0.2.0 (api 名有修改)
+
+- 更改 SchemaForm 组件的 valueName 为 valueBaseName
+- 修复 valueBaseName 有值时, valueType 为 dependency 时 columns 函数里 dataIndex 要处理 valueBaseName 值